Arel::Nodes::Conflict
