+++ lib/arel/extensions.rb
@@ -48,6 +48,7 @@ require 'arel/extensions/overlap'
 require 'arel/extensions/select_statement'
 require 'arel/extensions/insert_statement'
 require 'arel/extensions/default_values'
+require 'arel/extensions/conflict'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
